containerbuild: Add containerbuild-regionsrv
This provides a small service can be used by container building tools
like "docker" or "podman build" to obtain the required info to be able
to install packages inside the container image using the Public Cloud
RMTs. It is susposed to be used in conjunction with "container-suseconnect"
when building images inside on-demand public cloud instances.

The listen IP address and Port can be set via the CONTAINER_BUILD_IP and
CONTAINER_BUILD_PORT environment variables.