Refactor plugin load and region server arg generation
We will need to load the framework specific plugin and generate the region
server args in different contexts. One context is the existing addition of
the regionHint during registration. The other is to facilitate detection of
image movement when a user generates a "golden image" from a running
instance without cleaning up.