Server access testing expansion
At present when an IPv6 address is present in the configuration received from
the region server but the client network has no ipv6 capabilities the
detection if he server is responsive fails as the ipv4 access is not checked.
This update expands the code to check access via ipv4 shold the ipv6 access
fail.