Unhandeled Exception with IPv4 only client with IPv6 update server config
When the region server returns an update server configuration with IPv4 and
IPv6 but the client only has IPv4 capability an unhandeld exception is
triggered due to an uninitialized variable.

Once the unitialized variable issue is fixed, the code will write the server
cert to the trust anchors only with the IPv6 name as reference. This
is confusing. The updated code writes the cert with botht the IPv4 and IPv6
names.

Add tests for the updates smt.py implementation.