Update to 10.1.5 follow up fixes to IPv6
When the region server returns an update server configuration with IPv4
and IPv6 but the client only has IPv4 capability an unhandeld exception
is triggered due to an uninitialized variable.

Once the unitialized variable issue is fixed, the code will write the
server cert to the trust anchors only with the IPv6 name as reference.
This is confusing. The updated code writes the cert with both the IPv4
and IPv6 names.

In an IPv6 only setup no metadata could be retieved via the EC2 plugin. Try
to access the IPv6 endpoint for IMDS when IPv4 access fails.

Add tests