Fix failover path
In the failover path when determining an equivalent target server an attempt
is made to verify that the existing system credentials are valid on the new
target update server. This connection is made using the FQDN. However, at
the time the verification is made the resolution of the FQDN had not been
switched to the new IP address as such we attempted again to reach the server
that was detected in a down condition. The fix is to updated /etc/hosts
prior to attempting password verification.

Fix unit tests

Some unit tests were bleeding out of the test environment and dependent on the
state of the system where the tests were being run. Update the location of
the certs. These changes doe not yet address all issues around bleeding to
the test system.