Update the container registry code path
Change the code that runs the container registry registration process
to check if the registry has already been setup. At the moment
there is no check and the only reason why it's not called
multiple times are the early exit conditions which we have
to modify for the LTSS use cases.