Update the container registry code path
Change the code that runs the container registry registration
process to check if the registry has already been setup. At the
moment there is no check and the only reason why it's not called
multiple times was the early exit condition which prevented the
code from being executed. However with LTSS we need to support
the update case, thus any service feature (repo-setup, registry-setup
and ltss-registration) must be able to check if it was already
setup individually.