Added SLE12 ports fix to git
The required SLE12 fixes such that cloud-regionsrv-client
also functions on SLE12 regarding ipv6 only exists as a
patch file in the obs package sources. I suggest to add
and maintain the patch in git in the same way as we also
handle the .spec file. Otherwise there is a disconnect
in the git sources between the .spec and the patch files
it references