Deduplicate code that calls SUSEConnect
Instead of duplicating the code to prepare for a SUSEConnect
(or the respective transactional) command, provide a common
method named run_SUSEConnect which only receives the relevant
parameters. Apply the new method to the main registration
call as well as to the module registration. This refactor
was done also in preparation to an additional LTSS regcode
process which requires another SUSEConnect call.