Fix sequence order for initial registry setup
As part of the LTSS changes in the registration client the
initial setup of the registry moved before the base product
setup. This was considered not an issue but it is one. That's
because the setup of the registry reads the RMT credentials
from /etc/zypp/credentials.d/SCCcredentials which only exists
after the initial base product registration. This means the
initial registry setup has to follow the initial base
product registration. This commit fixes it