Fixed setup of registry hosts entry
There is a logic error in the code that sets up the hosts
entry. One method named add_hosts_entry() is responsible
for setting up the RMT and the registry hosts entries.
Both entries targets the same server IP and can only be
differentiated by their name and not by their IP. Next to
this design there is a method named has_rmt_in_hosts()
which checks by the IP of the server if the hosts entry
exists or not. Obviously this condition check is not
sufficient if there is the need to check individually for
the hosts entry of the RMT target or the registry target.

At the time we added support for the container registry
target, we require a more specific host entry checking
which is specific to the target system we want to add the
hosts entry for. This is because there are instances
which can run an old version of cloud-regionsrv-client
that only has support for the RMT target and therefore
has added a proper hosts entry for this target. If a customer
now updates cloud-regionsrv-client, the next reboot or
manual call for registercloudguest adds support for the
registry but did not update the hosts file because due
to the checking above the code thought everything is ok
because the IP based check for the hosts entry said
everything is good.

To fix this I changed the way how has_rmt_in_hosts()
works by using the new get_hosts() interface as well as
checking for the FQDN for RMT and not the IP. In addition I added
a new method called has_registry_in_hosts() which checks
for the FQDN for the registry. has_registry_in_hosts() is
then used to make sure a hosts entry for the registry
exists independently of the case that there is a RMT
hosts entry already present. This Fixes #195