Handle process structure change
During migration the various process get spawned. When looking for the main
zypper executable (Zypp-main) we also get information about the child
process. However we are only interested in the main process PID. Change the
logic to extract the PID such the we only look at the PID for the main
process.

Include testing of product migration in the integration test process. The
necessary changes for the zypper pid extraction did not show up in older
code bases and only manifested themselves during migration. As such we
have code sensitivity that was previously unkown. We want to catch this
during integration testing.

Add test method to simulate the changed output from the ps command due
to a changed process structure.