Make network access check message less alarming
At present when we test for network access by IP protocol version the
message in the log file makes it hard to differentiate whether this is
hard error or for information. While the message is set as info
it still contains the word "error" in the message. Make the message
less alarming in an attempt to not train users to ignore errors.

Enhance the testing of the function checking the network access. Previously
only the happy path was tested.

Write the error message to STDERR in case the user provides us with an
incorrectly formated address