Make network access check message less alarming
At present when we test for network access by IP protocol version the
message in the log file makes it hard to differentiate whether this is
a hard error or a message for information. While the message is set as info
it still contains the word "error". Make the message
less alarming in an attempt to not train users to ignore errors.

Enhance the testing of the function checking the network access. Previously
only the happy path was tested.

When the user provides a target update server IP and the ip cannot be
validated write the message to STDERR