package build update
- The way Python is being built has changed for SUSE distributions. Python 3.6
is, by design no longer available in products SLE Micro and later. We need
to use macros that are defined in the build service to pick up the interpreter
we want to use for building our code and to find the proper Python module
packages we depend on.
- Add a previoulsy forgotten patch for the SLE 12 build that we need
to maintain until the EOL of that code stream.