Refactor podman registry setup
Update __set_registries_conf_podman. The former implementation
had an index check problem if the suse registry was first in
the list. In addition it handles an in my eyes unneded check
for the insecure field of the private registry setup which is
never done by us in an insecure setup, insecure is always false.
Furthermore we should not change the registry setup if we can
find our entries inside but a user has done manual changes to
the order. I believe the proposed implementation also increases
the readability and logic flow of the code.