Refactor docker registry setup
Update __set_registries_conf_docker. The former implementation
had an index check problem if the suse registry was first in
the list. In addition we should not change the registry setup
if we can find our entries inside but a user has done manual
changes to the order. I believe the proposed implementation also
increases the readability and logic flow of the code.