Tidy up transactional-update integration
The transactional-update command integration was done inside the
run_SUSEConnect function as such creating a naming mismatch since the
function no longer only called SUSEConnect. Rename the function to a
generic name register_product and factor the detection about use
of the transactional-update command out into a separate function to
be re-used.

Co-authored-by: Marcus Schäfer <marcus.schaefer@suse.com>
Co-authored-by: Jesus Bermudez Velazquez <jesusbv@suse.com>