Use the region endpoint of IMDS in EC2
A while back EC2 IMDS gained a "region" endpoint to determine the region
in which the instance is running. Using this endpoint directly allows
us to drop the code that determined the region based on the information
from the "availability-zone" end point. This simplifies the code and testing.
Further this is more robust should AWS expand the zone naming as has
happened in the past.