Make sure to require container toolchain
cloud-regionsrv-client modifies config files from the
container toolchain, docker and podman. These files are
provided by packages under certain conditions, e.g.
permission and ownership. If the files are not present
on the system when registercloudguest is called they
are created under the umask() of the calling user which
is not neccessarily correct. Because of that, this
commit adds a requirement to the respective packages
providing the config files such that any modifications
retains the permission and ownership.