Update patch files to be clean again
Fix patch files such that they apply clean and without
producing a .orig file. Anytime a patch gets applied unclean
it leaves a .orig file and the package build fails because
of a new and unpackaged file. This commit fixes the patch
files to apply clean again and also makes them git compliant