Fix build with applied patches
The %patch macro does not support to call patch such that it
does not create .orig files. Under certain conditions patch
creates them and this will break the build for files found
but not packaged. From my perspective a successfully applied
patch even with an offset drift is still acceptable and
doesn't have to break the package build process. The way
how we maintain the patches for SLE12 will always provide
the offset drift as we are constantly working on other
parts of the code. If the patch applies is checked via
a github action per PR. Thus I believe the proposed patch
should be ok.