No need to for a versioned tarball
The way the source tarball is created produces a versioned
source tree inside of the tarball. There is no need to duplicate
this information in the file name of the tarfile. Providing
a consistent source file name prevents issues with package
maintenance in a way that on each version bump a new source file
needs to be deleted/added and also prevents issues with
multiple source files still available if the packager forgot
to delete the old one. Also spec file inconsistencies if only
a new source tarball gets committed but the spec and the old
source was forgotten to be updated/removed will no longer
occur.