Clean up when package gets removed
At present when the package gets removed we leave repositories behind that
point to the plugin that is supplied with this package. However with the
removal of this package the plugin is also removed. This leaves the system
in a state that is difficult to recover from and requires manual clean up.
We clean up the registration when the package is removed to avoid this
problem.