Turn docker into an optional setup
The simple docker requirement in the spec file of the
cloud-regionsrv-client package has turned into a big problem
for customers who don't want docker on their systems. This
commit makes docker a recommended package and refactors
the registration setup code in a way that docker becomes
an optional part of the setup procedure. With each call of
registercloudguest a check is made to see if docker is
installed. If it is installed the current setup is updated
with a docker related configuration if not already done