Fixed spec file %preun section
Latest spec change broke the spec file because there was
already a %preun section specified. Adding another one for
the same package is not supported and failed on rpm parsing.
This commit fixes the spec file and also fixes the invocation
for registercloudguest to be done only in an environment
providing a client configuration which is mandatory for the
tool to run. If no client config exists the package build
also fails because rpmbuild invokes the scripts at build
time.