Properly handle the root argument for zypper
When parsing the zypper command line to determine if we operate on a
different root system we did not account for the space in the argument
list to indicate the directory. This resulted in incorrect parsing results
for packages with "-R" in the package name. Resulting in failed package
updates.