Fix code state
Code was not releasable due to untested changes being merged triggering an
error when we were checking for access to the network for the IPv6 protocol.
One error was that IPAdress objects were passed for socket creation instead
of a string and the other error was that we did not pass an address at all.