Fix rfc3986 IP address check
Due to the change from IPv6Address object to string, the
check for rfc3986 was skipped because the isinstance call
no longer returns the expected object type. Fortunately the
coverage check uncovered the problem. This commit fixes
the code such that the rfc3986 check for the URL is active
again. On a personal note: isinstance() should be avoided.
According to object oriented coding practices something is
wrong with code that doesn't know what data type it deals
with.