Refactor state file setting
We have a number of state files that get added and removed based on the state
and progress of registration or based on billing model. Each of these
had it's own implementation for creating and removing the state file. Create
a utility functions to create and remove a state file and change the existing
code to use the new functions.