Modify the credentials_files_are_equal interface
Make the function less expensive on re-use. Previously the function
encapsulated the extraction of the root directory from the zypper
command line, meaning for every invocation we ended up calling the
ps utility. A use case has arisen where we want to re-use the function
to find orphaned credetials. In this case zypper is not running and
trying to extract the root path becomes and expensive undertaking.
With this change the function takes the rtarget root path as an
additional argument, which by default is the system root, i.e. /.
This pushed the responsibility of getting the target root to the caller.
At this point only the urlresolver needs the target root functionality.
Further, the urlresolver already has a feature to reduce the redundant
calls when we are in the same zypper process leading to additional
efficiency when extracting the target root.

Create 2 new constants to reduce repetition of the base credentials name
used as a string and the zypper credentials path.