Code cleanliness
Update the uses of os.path.join to use either os.sep.join or
os.path.normpath with os.sep.join to avoid oddities of th path.join
implementation if the second argument starts with a /. While we have not
had problems with this so far the previous use of path.join required
careful setu of constants and processing when files are being used. Also
discovered 1 incorrect use of path.join() and found another case where
string concatenation was used.