Move project setup to poetry
Update structure of the project to match with current
python standards. Use poetry to manage dependencies,
entrypoints and the setup of a development env as well
as the creation of the sdist tarball. Delete old, broken
and obsolete use of setuptools, setup.py and friends.
Stay compatible with runtime code down to SLE12.