Do not log exec call from register_product
The register_product method calls SUSEConnect with potential
sensitive information which should not be part of the log
file. Therefore the actual exec call is not logged by the
exec_subprocess method. However, a special log_information
variable is provided which obfuscates sensitive information
and sends the result as general information logging information
also ending up in the logfile