Refactor cleanup code
This commit refactors the cleanup code into single
responsibility methods for each type of cleanup. The
registration client has three areas for the cleanup

1. API delete requests to the SCC/RMT API
2. Cleanup of registration client specific cache data
3. Cleanup of configuration data created and or modified in /etc

The code changes refactors the existing cleanup methods
to not intermix the different cleanup types and provides
methods to implement cleanup methods properly to the
type of cleanup.