Refactor cleanup code
This commit refactors the cleanup code into single
responsibility methods for each type of cleanup. The
registration client has three areas for the cleanup

1. API delete requests to the SCC/RMT API
2. Cleanup of registration client specific cache data
3. Cleanup of configuration data created and or modified in /etc

The code changes refactors the existing cleanup methods
to not intermix the different cleanup types and provides
methods to implement cleanup methods properly to the
type of cleanup.

As an additional feature a new class named Git is provided
which allows to manage the cleanup of type 3 data (etc)
in git. This feature can be enabled via the --git option
and requires the user to install git beforehand as we
cannot force users to install git.

The git feature is handy for debugging and full control
over the data that the client really touches when performing
the registration. It's possible to switch between git and
no git managed data at any time. Customers who are not
against git to be present on a server and see the positive
side effects to have full control over data changed on
the server by a registration client also might find this
feature useful.

Last but not least the git feature would solve the still
present cleanup bug which prevents credentials files from
being cleaned up under certain conditions. As such we
still have to create a fix for this bug in the standard/legacy
cleanup code. With the refactoring code as part of this
PR this task will also become easier because the
cleanup methods are now more clear in structure and
processing.

Finally a number of unit tests got updated to be more
readable, less fragile and less painful to update in case
of code changes. There are still a number of tests that
could be improved though.