Fix Makefile package target
The Makefile looked up the version information by calling a
small python script that imports the version information
from source. This required to lookup the python interpreter
from the host which can be avoided now that we use poetry
for development. This commit moves the creation of the
version variable to the target where it is needed and
uses "poetry run python" to obtain the version information.