Add new Git class for content management
As of today all files created or modified by the client
registration in the systems /etc path are more or less
unknown. There is no way to differentiate between
modifications done by the registration client and changes
done by a user. As we are touching files that are likely
to be changed by users e.g. etc/hosts it is handy to put
all those under git control. This allows the client to
perform the following actions:

* git based restore to the state prior registration on cleanup
* possible detection of file changes done by a user between
  registration and cleanup
* proper rollback in case of exceptions or interrupts during
  the registration via python context manager

This feature can be enabled via the --git option
and requires the user to install git beforehand as we
cannot force users to install git. This Fixes #271