Fix credentials cleanup when registration fails
Based on the refactored cleanup code this commit fixes
the missing credentials and hosts cleanup when suseconnect
has failed. In addition the handling for NCCcredentials in
the client code was dropped. A %post section in the package
spec file was added which renames NCCcredentials to SCCcredentials
if applicable. Consequently the client implementation can
then be simplified to apply a handling for SCCcredentials only.