Update message regarding registry setup
The previous message was considered to contain insufficient information
and potentially leading to confusion, bsc#1254960. Update the message
to be more explicit about the meaning of "session". Also removes an exit
condition if docker setup fails. Even if the docker setup fails we should
try to setup the podman files.