Consolidate logging messages
Several logging messages are considered processing
instructions helpful in the log file but not of great
help on the console especially on successful registrations.
This commit turns a number of message to be debug
information which are listed precisely also with their
entry point in code in the log file but not on the
console. This Fixes #283